Ignore response when sending CRC request to trigger close
When using the `crcObject()` function to trigger a close event, the
`DfuTransportPrn` sets a `waitingForPacket` state. If the port is closed
after sending a CRC request, but before receiving a response, then the
`waitingForPacket` state is never cleared. Since the same transport
instance is used for multiple DFU updates, e.g. SoftDevice +
application, it will then still have the `waitingForPacket` state when
starting the next DFU update. In that case, the DFU will fail with "DFU
transport tried to read() while another read() was still waiting".

Since we are not interested in the response to these CRC requests, we
can avoid the issue by just writing the CRC command without waiting for
a response.